requirements.txt
openai
httpx
praw
fastapi
uvicorn[standard]
redis
asyncpg
sqlalchemy[asyncio]
alembic
pydantic-settings
python-dotenv
apscheduler
pandas
numpy
beautifulsoup4
feedparser
aiohttp
websockets
yfinance
pytrends
discord.py
telethon
apify-client
lxml
